Adding middleware, fix several issues

+++ UIvent BE/src/controllers/event.controller.js
@@ -20,6 +20,16 @@ exports.createEvent = async (req, res) => {
         return baseResponse(res, false, 400, "Please enter the required fields", null);
     }
 
+    if (req.body.organizer_id !== req.user.organization && req.user.role !== "admin") {
+        return baseResponse(
+            res, 
+            false, 
+            403, 
+            "You are not authorized to create an event for this organization", 
+            null
+        );
+    }
+
     const {start_date, start_time, end_date, end_time} = req.body;
 
     const start_datetime = `${start_date}T${start_time}:00Z`;
@@ -102,18 +112,29 @@ exports.updateEvent = async (req, res) => {
         return baseResponse(res, false, 400, "Please enter the required fields", null);
     }
 
+    let event = await eventRepository.getEventById(req.params.id);
+
+    if (!event) {
+            return baseResponse(res, false, 404, "Event not found", null);
+    }
+
+    if (event.organizer_id !== req.user.organization && req.user.role !== "admin") {
+        return baseResponse(
+            res, 
+            false, 
+            403, 
+            "You are not authorized to update an event for this organization", 
+            null
+        );
+    }
+
     const {start_date, start_time, end_date, end_time} = req.body;
 
     const start_datetime = `${start_date}T${start_time}:00Z`;
     const end_datetime = `${end_date}T${end_time}:00Z`;
 
     try {
-        const event = await eventRepository.updateEvent(req.params.id, { start_datetime, end_datetime }, req.body);
-
-        if (!event) {
-            return baseResponse(res, false, 404, "Event not found", null);
-        }
-
+        event = await eventRepository.updateEvent(req.params.id, { start_datetime, end_datetime }, req.body);
         return baseResponse(res, true, 200, "Event updated successfully", event);
     }
     catch (error) {
@@ -123,13 +144,24 @@ exports.updateEvent = async (req, res) => {
 };
 
 exports.deleteEvent = async (req, res) => {
-    try {
-        const event = await eventRepository.deleteEvent(req.params.id);
+    let event = await eventRepository.getEventById(req.params.id);
 
-        if (!event) {
+    if (!event) {
             return baseResponse(res, false, 404, "Event not found", null);
-        }
+    }
 
+    if (event.organizer_id !== req.user.organization && req.user.role !== "admin") {
+        return baseResponse(
+            res, 
+            false,
+            403, 
+            "You are not authorized to delete an event for this organization", 
+            null
+        );
+    }
+    
+    try {
+        event = await eventRepository.deleteEvent(req.params.id);
         return baseResponse(res, true, 200, "Event deleted successfully", event);
     }
     catch (error) {
